fix: cambia los route de user al de store y viceversa
