+++ src/app/api/profile/store/route.ts
@@ -2,15 +2,17 @@ import { getTokenFromCookie } from "@/lib/getTokenFromCookie";
 import { ResourceNotFound, UnauthorizedError } from "@/lib/errors/customErrors";
 import { getUserFromToken } from "@/lib/getUserFromToken";
 import { handleError } from "@/lib/errors/errorHandler";
-import { getUserProfile, updateUser } from "@/features/users/user.service";
+import { getUserProfile, updateStore } from "@/features/users/user.service";
 import { successResponse } from "@/lib/responses";
 import { ProfileData } from "@/features/users/user.dto";
 import { NextRequest} from "next/server";
+
 
 export async function PATCH(req: NextRequest) {
   try {
     const token = getTokenFromCookie(req);
 
+    // ?: Se está validando la validez del token?
     if (!token) {
       throw new UnauthorizedError('Token inválido o faltante');
     }
@@ -27,10 +29,11 @@ export async function PATCH(req: NextRequest) {
       throw new ResourceNotFound('Usuario no encontrado');
     }
 
-    const { store, ...userEntity } = profile;
+    const { store: storeEntity } = profile;
 
-    const { password, userRole, ...safeUser } = await updateUser(userEntity, body);
-    const updatedProfile: ProfileData = { ...safeUser, store }
+    const updatedStore = await updateStore(storeEntity, body);
+    
+    const updatedProfile: ProfileData = { ...profile, store: updatedStore }
 
     return successResponse(updatedProfile);
   } catch (error) {
