+++ src/app/api/profile/user/route.ts
@@ -1,18 +1,17 @@
+
 import { getTokenFromCookie } from "@/lib/getTokenFromCookie";
 import { ResourceNotFound, UnauthorizedError } from "@/lib/errors/customErrors";
 import { getUserFromToken } from "@/lib/getUserFromToken";
 import { handleError } from "@/lib/errors/errorHandler";
-import { getUserProfile, updateStore } from "@/features/users/user.service";
+import { getUserProfile, updateUser } from "@/features/users/user.service";
 import { successResponse } from "@/lib/responses";
 import { ProfileData } from "@/features/users/user.dto";
 import { NextRequest} from "next/server";
-
 
 export async function PATCH(req: NextRequest) {
   try {
     const token = getTokenFromCookie(req);
 
-    // ?: Se está validando la validez del token?
     if (!token) {
       throw new UnauthorizedError('Token inválido o faltante');
     }
@@ -29,11 +28,10 @@ export async function PATCH(req: NextRequest) {
       throw new ResourceNotFound('Usuario no encontrado');
     }
 
-    const { store: storeEntity } = profile;
+    const { store, ...userEntity } = profile;
 
-    const updatedStore = await updateStore(storeEntity, body);
-    
-    const updatedProfile: ProfileData = { ...profile, store: updatedStore }
+    const { password, userRole, ...safeUser } = await updateUser(userEntity, body);
+    const updatedProfile: ProfileData = { ...safeUser, store }
 
     return successResponse(updatedProfile);
   } catch (error) {
